feat(store): improve createSelector function typings
