+++ packages/store/src/public_api.ts
@@ -41,7 +41,6 @@ export {
 } from './symbols';
 export { Selector } from './decorators/selector/selector';
 export { getActionTypeFromInstance, actionMatcher } from './utils/utils';
-export { createSelector } from './selectors/selector-utils';
 export { NgxsExecutionStrategy } from './execution/symbols';
 export { ActionType, ActionOptions } from './actions/symbols';
 export { NoopNgxsExecutionStrategy } from './execution/noop-ngxs-execution-strategy';
@@ -55,6 +54,7 @@ export {
   createModelSelector,
   createPickSelector,
   createPropertySelectors,
+  createSelector,
   PropertySelectors,
   TypedSelector,
 } from './selectors';
