+++ packages/store/src/selectors/index.ts
@@ -1,4 +1,5 @@
 export * from './create-model-selector';
 export * from './create-pick-selector';
 export * from './create-property-selectors';
+export * from './create-selector';
 export * from './selector-types.util';
